@@ -14,49 +14,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
-    "@angular/http": "^6.1.10",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.1.10",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@ngx-translate/core": "^10.0.2",
     "@ngx-translate/http-loader": "^3.0.1",
-    "@rocketshow/designer": "^1.29.0",
-    "angular-sortablejs": "^2.7.0",
+    "@rocketshow/designer": "^1.31.0",
     "angular2-websocket": "^0.9.7",
     "bootstrap": "^4.1.2",
     "core-js": "^2.5.7",
-    "d3": "^5.6.0",
     "file-saver": "^1.3.8",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.3.1",
     "ngx-bootstrap": "^3.1.3",
     "ngx-bootstrap-slider": "^1.6.0",
     "ngx-dropzone-wrapper": "^4.9.0",
+    "ngx-sortablejs": "^11.1.0",
     "ngx-toastr": "^8.10.2",
     "ngx-translate-multi-http-loader": "^3.0.0",
-    "rxjs": "^6.2.2",
+    "rxjs": "^6.6.3",
     "sortablejs": "^1.9.0",
     "split.js": "^1.5.10",
     "three": "^0.104.0",
+    "tslib": "^1.9.0",
     "wavesurfer": "^1.3.4",
     "wavesurfer.js": "^3.0.0",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.26"
+    "web-animations-js": "^2.3.2",
+    "web-audio-beat-detector": "^8.1.1",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^6.2.9",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/language-service": "^6.1.10",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "^2.0.3",
     "@types/node": "^6.0.114",
-    "codelyzer": "4.3.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "^1.7.1",
@@ -71,6 +70,6 @@
     "rmdir-cli": "^2.0.6",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "^2.8.1"
+    "typescript": "^3.2.4"
   }
 }
